Disable pickups for non-active stores
